# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: CraftForge Web API
  version: 1.0.0
  description: |
    Especificacion ejecutable del backend V1 para CraftForge Web.
servers:
  - url: http://localhost:3000
    description: local
...
paths:
...
  /api/v1/modpacks/{modpackId}/exports:
    post:
      tags: [Exports]
      summary: Encola exportacion de modpack
      parameters:
        - $ref: "#/components/parameters/ModpackId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExportRequest"
      responses:
        "202":
          description: Job encolado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ExportJob"
                required: [data]
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/exports/{jobId}:
    get:
      tags: [Exports]
      summary: Consulta estado de exportacion
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Estado del job
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ExportJob"
                required: [data]
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/exports/{jobId}/download:
    get:
      tags: [Exports]
      summary: Descarga ZIP cuando job este completado
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "302":
          description: Redirige a URL de descarga local
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Job aun no completado
components:
  parameters:
    ModpackId:
      in: path
      name: modpackId
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Request invalido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    CreateExportRequest:
      type: object
      properties:
        target:
          type: string
          enum: [CLIENT, SERVER, BOTH]
      required: [target]
    ExportJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        modpackId:
          type: integer
        target:
          type: string
          enum: [CLIENT, SERVER, BOTH]
        status:
          type: string
          enum: [queued, running, completed, failed]
        outputPath:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, modpackId, target, status, createdAt]